+++ test/instream/connection/config_test.exs
@@ -5,16 +5,6 @@ defmodule Instream.Connection.ConfigTest do
   alias Instream.TestHelpers.Connections.DefaultConnection
   alias Instream.TestHelpers.Connections.EnvConnection
 
-  test "missing configuration raises", %{test: test} do
-    exception =
-      assert_raise ArgumentError, fn ->
-        Config.validate!(test, __MODULE__)
-      end
-
-    assert String.contains?(exception.message, inspect(__MODULE__))
-    assert String.contains?(exception.message, inspect(test))
-  end
-
   test "runtime configuration changes", %{test: test} do
     conn = Module.concat([__MODULE__, RuntimeChanges])
     key = :runtime_testing_key